feat(observability): split metrics + inspect wire types into SDK-free leaf packages
Anchor #2 of the cmd/api binary-size reduction (reliable#2153 / #2141):
reliable's serverless function is at the 250 MB hard limit, held there by
the AWS + GCP SDK service packages imported transitively via the
observability inspect/metrics paths. To let reliable become a thin proxy
that deserializes ui-core's responses *without* importing those clients,
the pure-data result/wire types now live in dedicated SDK-free leaf
packages:

- pkg/observability/metrics/metricstypes: MetricsResult, ResourceMetrics,
  MetricSeries, Datapoint, MetricsFilter, ResourceID.
- pkg/observability/inspect/inspecttypes: SubRequest, SubResult,
  BatchRequest, BatchResponse, MaxBatchSubs.

The parent metrics / inspect packages re-export every type as a Go type
alias, so all existing in-tree callers (and the wire shape) are
unchanged. Each leaf package ships a `go list -deps` poka-yoke test that
fails if any aws-sdk-go / cloud.google.com/go / google.golang.org/api
dependency ever leaks back in.
