+++ pkg/observability/inspect/inspecttypes/types.go
@@ -1,0 +1,89 @@
+// Package inspecttypes holds the SDK-free wire-envelope types for the
+// inspect batch surface (SubRequest, SubResult, BatchRequest,
+// BatchResponse) plus the MaxBatchSubs cap. It is a leaf package with
+// ZERO cloud-SDK imports so a consumer that only needs to (de)serialize
+// the inspect wire shape — e.g. luthersystems/reliable's thin proxy
+// after the cloud reads moved into ui-core (reliable#2153) — can do so
+// without dragging the AWS / GCP SDK clients pulled in by the parent
+// inspect package's Dispatcher into its binary.
+//
+// The parent pkg/observability/inspect package re-exports every type
+// here as a Go type alias, so existing callers of inspect.SubResult etc.
+// are unaffected and the wire shape is byte-for-byte identical.
+//
+// JSON shapes are wire-stable. Changing a field name or json tag is a
+// wire-breaking change and requires coordinated rollout across the
+// reliable handler, the ui-core /observability surface, and every
+// MCP-server callsite. Tests in inspect/types_test.go pin the exact
+// json tags and `omitempty` semantics so drift surfaces at unit-test
+// time rather than at HTTP-decode time.
+package inspecttypes
+
+// MaxBatchSubs caps the number of sub-probes per batch request. The
+// dispatcher rejects requests with len(Subs) > MaxBatchSubs. The number
+// is wire-coupled — both the client (MCP server / reliable HTTP handler)
+// and the server (batch dispatcher) read this same constant, so changing
+// it requires coordinated rollout.
+const MaxBatchSubs = 32
+
+// SubRequest is one probe within a batch. Service is a registry name
+// recognized by pkg/observability/service_actions.go's AWSServiceNames
+// or GCPServiceNames; Action is a registry-defined verb. Filters carries
+// arbitrary JSON the per-service handler interprets — left empty for
+// actions that need no filter.
+//
+// Detail / raw flags are deliberately NOT exposed: batch always returns
+// summarized results. Callers needing detail or raw output should use
+// the singular awsinspect / gcpinspect tools.
+//
+// The `jsonschema:"..."` tags are read by MCP clients via
+// github.com/google/jsonschema-go reflection to publish per-property
+// descriptions on tools/list. They are inert for encoding/json.
+type SubRequest struct {
+	Service string `json:"service" jsonschema:"Cloud service to query (e.g. 'ec2', 'rds', 's3' for AWS; 'compute', 'storage', 'cloudsql' for GCP). Use the singular awsinspect/gcpinspect tool with action='list-actions' to discover supported services."`
+	Action  string `json:"action" jsonschema:"Operation on the service (e.g. 'describe-instances', 'list-buckets', 'list-actions' for discovery, 'list-metrics' / 'get-metrics' for time-series)."`
+	Filters string `json:"filters,omitempty" jsonschema:"Optional JSON-encoded filter object passed through to the underlying API. Examples: '{\"hours\":6}' for metric windows, '{\"days\":7}' for cost queries."`
+}
+
+// SubResult is one probe's outcome. Index pins the result back to the
+// SubRequest at the same index in the original Subs slice — the fan-out
+// dispatcher may complete out of order. OK is true iff Error is empty;
+// Result is set iff OK is true; Error is set iff OK is false. DurationMS
+// captures the per-probe latency in milliseconds and is always emitted
+// (even when zero) for observability.
+type SubResult struct {
+	Index      int    `json:"index"`
+	Service    string `json:"service"`
+	Action     string `json:"action"`
+	OK         bool   `json:"ok"`
+	Result     any    `json:"result,omitempty"`
+	Error      string `json:"error,omitempty"`
+	DurationMS int64  `json:"duration_ms"`
+}
+
+// BatchRequest is the wire envelope for both
+// POST /api/v2/aws/inspect/batch and POST /api/v2/gcp/inspect/batch.
+// SessionID identifies the calling reliable session; the dispatcher
+// resolves it to credentials + project context.
+//
+// A single shape covers both clouds because the request shape is
+// cloud-agnostic — the route the request lands on (`/aws/...` vs
+// `/gcp/...`) carries the cloud selection.
+type BatchRequest struct {
+	SessionID string       `json:"session_id"`
+	Subs      []SubRequest `json:"subs"`
+}
+
+// BatchResponse is the wire envelope returned by the same two endpoints.
+// OK is the HTTP-envelope success bit — true iff the dispatcher itself
+// ran to completion (HTTP 200 path). Per-sub success/failure is encoded
+// in Results[i].OK and Results[i].Error, NOT aggregated into this outer
+// OK: a partial-failure batch (some Results[i].OK == false) still
+// returns outer OK == true.
+//
+// Results is index-aligned with the original Subs slice:
+// Results[i].Index == i.
+type BatchResponse struct {
+	OK      bool        `json:"ok"`
+	Results []SubResult `json:"results"`
+}
